requirements.txt
Flask==2.3.0
SQLAlchemy==2.0.0
cryptography==39.0.1
argon2-cffi==21.3.0
flask-limiter==2.3.1
